file reorganization and build break fixes
