--- packages/framework/theming-react-native/src/NativeModule/ThemingModule.native.ts
@@ -1,4 +1,0 @@
-import { createThemingModuleHelper } from './ThemingModuleHelpers';
-import { getThemingModule } from '@fluentui-react-native/win32-theme';
-
-export const ThemingModuleHelper = createThemingModuleHelper(...getThemingModule());
