--- packages/framework/theming-react-native/src/NativeModule/__mocks__/ThemingModule.native.ts
@@ -1,3 +1,0 @@
-import { createMockThemingModuleHelper } from '../MockThemingModule';
-
-export const ThemingModuleHelper = createMockThemingModuleHelper();
