+++ packages/theming/win32-theme/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@fluentui-react-native/win32-theme",
   "version": "0.1.1",
-  "description": "Typing only package for fluentui-react-native theme types",
+  "description": "Win32 office theme that works with the theming native module",
   "repository": {
     "type": "git",
     "url": "https://github.com/microsoft/just"
