+++ packages/theming/win32-theme/src/NativeModule/_mocks_/enableNativeModule.native.ts
@@ -1,0 +1,3 @@
+export function enableNativeModule(): boolean {
+  return false;
+}
