+++ packages/theming/win32-theme/src/NativeModule/index.ts
@@ -1,0 +1,4 @@
+export * from './enableNativeModule';
+export * from './fallbackOfficeModule';
+export * from './getThemingModule';
+export * from './officeThemingModule';
