+++ packages/theming/win32-theme/src/index.ts
@@ -1,7 +1,4 @@
 export * from './createOfficeTheme';
 export * from './createPartialOfficeTheme';
-export * from './enableNativeModule';
-export * from './fallbackOfficeModule';
-export * from './getThemingModule';
-export * from './officeThemingModule';
+export * from './NativeModule/index';
 export * from './paletteFromOfficeColors';
